LayoutTests/platform/ios/imported/w3c/web-platform-tests/visual-viewport/viewport-scrollbars-cause-resize-in-iframe-expected.txt

Viewport: Scrollbars Cause Resize In IFrame

Test Description: This test checks that the appearance of classic scrollbars will cause a resize event to be fired at window.visualViewport in iframe.



FAIL the appearance of classic scrollbars will fire a resize eventat window.visualViewport in iframe assert_equals: expected 0 but got 1
